+++ src/def/osrs/npc_def.rs
@@ -6,16 +6,14 @@ use std::{
 
 use serde::{ Serialize, Deserialize };
 
-use crate::{ Definition, ext::ReadExt, util };
+use super::Definition;
+use crate::{ ext::ReadExt, util };
 
 /// Contains all the information about a certain npc fetched from the cache through
 /// the [NpcLoader](struct.NpcLoader.html).
-/// 
-/// The `NpcModelData` and the `NpcAnimationData` were hidden in the documents
-/// because these are rarely accessed, they contain useless information in most use-cases. 
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq, Debug, Default)]
 pub struct NpcDefinition {
-    pub id: u32,
+    pub id: u16,
     pub name: String,
     pub size: usize,
     pub actions: [String; 5],
@@ -62,7 +60,7 @@ pub struct NpcAnimationData {
 
 impl Definition for NpcDefinition {
     #[inline]
-    fn new(id: u32, buffer: &[u8]) -> io::Result<Self> {
+    fn new(id: u16, buffer: &[u8]) -> io::Result<Self> {
         let mut reader = BufReader::new(buffer);
         let npc_def = decode_buffer(id, &mut reader)?;
 
@@ -70,7 +68,7 @@ impl Definition for NpcDefinition {
     }
 }
 
-fn decode_buffer(id: u32, reader: &mut BufReader<&[u8]>) -> io::Result<NpcDefinition> {
+fn decode_buffer(id: u16, reader: &mut BufReader<&[u8]>) -> io::Result<NpcDefinition> {
     let mut npc_def = NpcDefinition {
         id,
         interactable: true,
